requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.7
python-multipart
stripe>=9.8
qrcode[pil]>=7.4
python-dotenv>=1.0
aiosqlite>=0.20
pydantic-settings>=2.1
loguru>=0.7
python-json-logger>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7